requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
flask==3.0.3
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
validators==0.33.0
werkzeug==3.0.3
wizart-vision==1.0.6
zipp==3.19.2
